requirements.txt
azure-functions
azure-core==1.30.0
azure-identity==1.15.0
azure-storage-blob==12.19.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
idna==3.6
isodate==0.6.1
msal==1.27.0
msal-extensions==1.1.0
packaging==23.2
portalocker==2.8.2
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
pywin32==306
requests==2.31.0
six==1.16.0
typing_extensions==4.10.0
urllib3==2.2.1
